Parse IShares list

parses the raw JSON data (usually returned as nested lists) into a clean dataframe of character data. We decided to return character data to avoid data type issues during the parsing. When the JSON data is downloaded using as = "parsed", the parsing is done using the functions in the https://tidyr.tidyverse.org/index.html package

parse_summary_data(res_parsed)

Arguments

res_parsed
a list returned by the httr::content function. The list ha the following attributes:

code

message

status

code

The most important attribute is data which contains the actual IShares data. In particular, it's a list with the following attributes:

tableData a list with two attributes: columns and data. The first one store the column names and the second the actual data

config.

Value

a dataframe of characters. Each row containd summary data for one etf